Extracted logic from destination server
